artistry-backend/advise/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
sentence-transformers==2.2.2
pymongo==4.6.0
pillow==10.1.0
numpy==1.24.4
accelerate==0.24.1